+++ lib/Model/SmartTransactionsPrepareDynamicDescriptor.php
@@ -2,16 +2,18 @@
 
 namespace Secuconnect\Client\Model;
 
+use \ArrayAccess;
+
 /**
- * SmartTransactionsContainer
+ * SmartTransactionsPrepareDynamicDescriptor
  *
  * @category Class
- * @description Payment instrument
+ * @description Dynamic Descriptor to allow the cardholder to recognize the business name and city on their card or bank statement.
  * @package  Secuconnect\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
-class SmartTransactionsContainer extends ProductInstanceUID
+class SmartTransactionsPrepareDynamicDescriptor implements ArrayAccess
 {
     const DISCRIMINATOR = null;
 
@@ -19,15 +21,15 @@ class SmartTransactionsContainer extends ProductInstanceUID
       * The original name of the model.
       * @var string
       */
-    protected static $swaggerModelName = 'SmartTransactionsContainer';
+    protected static $swaggerModelName = 'SmartTransactionsPrepareDynamicDescriptor';
 
     /**
       * Array of property to type mappings. Used for (de)serialization
       * @var string[]
       */
     protected static $swaggerTypes = [
-        'type' => 'string',
-        'token_status' => '\Secuconnect\Client\Model\PaymentContainerTokenStatus'
+        'merchant_name' => 'string',
+        'merchant_city' => 'string'
     ];
 
     /**
@@ -35,18 +37,18 @@ class SmartTransactionsContainer extends ProductInstanceUID
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'type' => null,
-        'token_status' => null
+        'merchant_name' => null,
+        'merchant_city' => null
     ];
 
     public static function swaggerTypes()
     {
-        return self::$swaggerTypes + parent::swaggerTypes();
+        return self::$swaggerTypes;
     }
 
     public static function swaggerFormats()
     {
-        return self::$swaggerFormats + parent::swaggerFormats();
+        return self::$swaggerFormats;
     }
 
     /**
@@ -54,8 +56,8 @@ class SmartTransactionsContainer extends ProductInstanceUID
      * @var string[]
      */
     protected static $attributeMap = [
-        'type' => 'type',
-        'token_status' => 'token_status'
+        'merchant_name' => 'merchant_name',
+        'merchant_city' => 'merchant_city'
     ];
 
     /**
@@ -63,8 +65,8 @@ class SmartTransactionsContainer extends ProductInstanceUID
      * @var string[]
      */
     protected static $setters = [
-        'type' => 'setType',
-        'token_status' => 'setTokenStatus'
+        'merchant_name' => 'setMerchantName',
+        'merchant_city' => 'setMerchantCity'
     ];
 
     /**
@@ -72,24 +74,30 @@ class SmartTransactionsContainer extends ProductInstanceUID
      * @var string[]
      */
     protected static $getters = [
-        'type' => 'getType',
-        'token_status' => 'getTokenStatus'
+        'merchant_name' => 'getMerchantName',
+        'merchant_city' => 'getMerchantCity'
     ];
 
     public static function attributeMap()
     {
-        return parent::attributeMap() + self::$attributeMap;
+        return self::$attributeMap;
     }
 
     public static function setters()
     {
-        return parent::setters() + self::$setters;
+        return self::$setters;
     }
 
     public static function getters()
     {
-        return parent::getters() + self::$getters;
+        return self::$getters;
     }
+
+    /**
+     * Associative array for storing property values
+     * @var mixed[]
+     */
+    protected $container = [];
 
     /**
      * Constructor
@@ -97,10 +105,8 @@ class SmartTransactionsContainer extends ProductInstanceUID
      */
     public function __construct(array $data = null)
     {
-        parent::__construct($data);
-
-        $this->container['type'] = isset($data['type']) ? $data['type'] : null;
-        $this->container['token_status'] = isset($data['token_status']) ? $data['token_status'] : null;
+        $this->container['merchant_name'] = isset($data['merchant_name']) ? $data['merchant_name'] : null;
+        $this->container['merchant_city'] = isset($data['merchant_city']) ? $data['merchant_city'] : null;
     }
 
     /**
@@ -110,7 +116,7 @@ class SmartTransactionsContainer extends ProductInstanceUID
      */
     public function listInvalidProperties()
     {
-        $invalid_properties = parent::listInvalidProperties();
+        $invalid_properties = [];
 
         return $invalid_properties;
     }
@@ -128,43 +134,43 @@ class SmartTransactionsContainer extends ProductInstanceUID
 
 
     /**
-     * Gets type
+     * Gets merchant_name
      * @return string
      */
-    public function getType()
+    public function getMerchantName()
     {
-        return $this->container['type'];
+        return $this->container['merchant_name'];
     }
 
     /**
-     * Sets type
-     * @param string $type The container type
+     * Sets merchant_name
+     * @param string $merchant_name Merchant Name to recognition on card or bank statement.  _Note: [1 - 22] characters_
      * @return $this
      */
-    public function setType($type)
+    public function setMerchantName($merchant_name)
     {
-        $this->container['type'] = $type;
+        $this->container['merchant_name'] = $merchant_name;
 
         return $this;
     }
 
     /**
-     * Gets token_status
-     * @return \Secuconnect\Client\Model\PaymentContainerTokenStatus
+     * Gets merchant_city
+     * @return string
      */
-    public function getTokenStatus()
+    public function getMerchantCity()
     {
-        return $this->container['token_status'];
+        return $this->container['merchant_city'];
     }
 
     /**
-     * Sets token_status
-     * @param \Secuconnect\Client\Model\PaymentContainerTokenStatus $token_status token_status
+     * Sets merchant_city
+     * @param string $merchant_city Merchant City to recognition on card or bank statement.  _Note: [1 - 13] characters_
      * @return $this
      */
-    public function setTokenStatus($token_status)
+    public function setMerchantCity($merchant_city)
     {
-        $this->container['token_status'] = $token_status;
+        $this->container['merchant_city'] = $merchant_city;
 
         return $this;
     }